requirements.txt
fastapi
uvicorn[standard]
asyncpg
SQLAlchemy[asyncio]
psycopg2-binary
python-dotenv
